04-02-2019

+++ InventoryManagement/Areas/Inventory/Models/Product.cs
@@ -11,7 +11,7 @@ namespace InventoryManagement.Areas.Inventory.Models
         [Key]
         public int ProductID { get; set; }
         public string ProductName { get; set; }
-        public decimal SalesPrice { get; set; }
+        public double SalesPrice { get; set; }
         public int ConcernID { get; set; }
         public string Description { get; set; }
         public DateTime CreationDate { get; set; }
@@ -19,7 +19,7 @@ namespace InventoryManagement.Areas.Inventory.Models
         public int IsDelete { get; set; }
         public int CategoryID { get; set; }
         public string ProductCode { get; set; }
-        public decimal PurchasePrice { get; set; }
+        public double PurchasePrice { get; set; }
         public decimal Quantity { get; set; }
         public int UnitID { get; set; }
     }
